Downloads/Telegram Desktop/B-A/Chain-BD/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cassandra-driver==3.28.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
requests==2.31.0
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0
xgboost==2.0.3
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2